+++ sentry-ruby/lib/sentry-ruby.rb
@@ -232,6 +232,7 @@ module Sentry
       yield(config) if block_given?
       config.detect_release
       apply_patches(config)
+      config.validate
       client = Client.new(config)
       scope = Scope.new(max_breadcrumbs: config.max_breadcrumbs)
       hub = Hub.new(client, scope)